Split BlockchainService test
